Maintain: Initialize various structures of x86-64.
[x86-64] Structure initialization.
    -- We don't have the desire to make x86-64 now, so we only distinguish
        between the two different data structures for compatible instructions.

[memory] Add and maintain.
    -- Simply added memory subsystem allocation support
        for pmd pud p4d page tables.

+++ drivers/dt/iterator.c
@@ -115,13 +115,13 @@ state dt_iterative_phandle_next(struct dt_phandle_iterator *iter)
 
     if (unlikely(iter->curr + count > iter->list_end)) {
         if (iter->cells_name)
-            pr_err("node %p %s = %d found %d\n",
+            pr_err("node %p %s = %d found %ld\n",
                     iter->parent, iter->node->name,
-                    count, iter->list_end - iter->curr);
+                    count, (long)(iter->list_end - iter->curr));
         else
-            pr_err("node %p phandle %s needs %d, found %d\n",
+            pr_err("node %p phandle %s needs %d, found %ld\n",
                     iter->parent, iter->node->name,
-                    count, iter->list_end - iter->curr);
+                    count, (long)(iter->list_end - iter->curr));
         goto error;
     }
